desktop/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
cryptography
grpcio
grpcio-tools
